containerd_protos::services::content

Module v1

Modules

content_client

Generated client implementations.

Structs

AbortRequest

DeleteContentRequest

Info

InfoRequest

InfoResponse

ListContentRequest

ListContentResponse

ListStatusesRequest

ListStatusesResponse

ReadContentRequest

ReadContentRequest defines the fields that make up a request to read a portion of data from a stored object.

ReadContentResponse

ReadContentResponse carries byte data for a read request.

Status

StatusRequest

StatusResponse

UpdateRequest

UpdateResponse

WriteContentRequest

WriteContentRequest writes data to the request ref at offset.

WriteContentResponse

WriteContentResponse is returned on the culmination of a write call.

Enums

WriteAction

WriteAction defines the behavior of a WriteRequest.