containerd_protos::services::content::v1

Enum WriteAction

#[repr(i32)]
pub enum WriteAction {
    Stat = 0,
    Write = 1,
    Commit = 2,
}

WriteAction defines the behavior of a WriteRequest.

Variants

Stat = 0

WriteActionStat instructs the writer to return the current status while holding the lock on the write.

Write = 1

WriteActionWrite sets the action for the write request to write data.

Any data included will be written at the provided offset. The transaction will be left open for further writes.

This is the default.

Commit = 2

WriteActionCommit will write any outstanding data in the message and commit the write, storing it under the digest.

This can be used in a single message to send the data, verify it and commit it.

This action will always terminate the write.

Implementations

impl WriteAction

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of WriteAction.

pub fn from_i32(value: i32) -> Option<WriteAction>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a WriteAction, or None if value is not a valid variant.

impl WriteAction

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<Self>

Creates an enum from field names used in the ProtoBuf definition.