containerd_protos::types

Module transfer

Structs

AuthRequest

AuthRequest is sent as a callback on a stream

AuthResponse

Data

ImageExportStream

ImageImportStream

ImageReference

ImageReference is used to create or find a reference for an image

ImageStore

OciRegistry

Progress

RegistryResolver

UnpackConfiguration

WindowUpdate

Enums

AuthType

HttpDebug